api-server/requirements.txt
fastapi[standard] 
python-dotenv 
uvicorn
pydantic-settings
google-generativeai


